VFB-311 separate reducer helper file into server and client side files
